requirements.txt
fastapi~=0.123.0
uvicorn~=0.38.0
pydantic~=2.11.7
python-dotenv~=1.1.1
google-generativeai
googleapis-common-protos~=1.70.0